Rename CategoryList
